divide up testing files for clarity
